added size and fixed up the player class and entity class

+++ zombieDodge/src/config.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <iostream>
+
 // Window related defines //
 const int WINDOW_WIDTH = 800;
 const int WINDOW_HEIGHT = 600;
@@ -7,4 +9,4 @@ const std::string WINDOW_CAPTION = "Stupid Zombies";
 
 // Game related defines //
 const int FRAMES_PER_SECOND= 30;
-const float FRAME_RATE = 1000/FRAMES_PER_SECOND;
+const double FRAME_RATE = 1000/FRAMES_PER_SECOND;
